pyproject.toml
[project]
name = "youtube-playlist-downloader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "charset-normalizer==3.4.5",
    "ffmpeg-python==0.2.0",
    "future==1.0.0",
    "g2p-en>=2.1.0",
    "jinja2==3.1.6",
    "llvmlite==0.46.0",
    "markupsafe==3.0.2",
    "more-itertools==10.8.0",
    "mpmath==1.3.0",
    "networkx==3.6.1",
    "numba==0.64.0",
    "openai-whisper==20250625",
    "python-dotenv>=1.0.0",
    "pytube==15.0.0",
    "requests==2.32.5",
    "sacremoses>=0.1.1",
    "sentencepiece>=0.2.1",
    "setuptools==70.2.0",
    "sympy==1.14.0",
    "tiktoken==0.12.0",
    "torch>=2.5.0",
    "transformers>=5.3.0",
    "urllib3==2.6.3",
    "youtube-dl==2021.12.17",
    "yt-dlp>=2026.3.3",
]
